+++ src/postman/actions/pm-environment-update.action.ts
@@ -2,6 +2,7 @@ import { PostmanEnvironment, PostmanEnvironmentMinMetadata } from '../api/types'
 import { IPMACAction } from './action.interface'
 import { PostmanAPI } from '../api'
 import { PMACWorkspace } from '../../file-system/types'
+import { PMACEnvironmentGetPMACMapAction } from './pmac-environment-get-map.action'
 
 export class PMEnvironmentUpdateAction implements IPMACAction<PostmanEnvironmentMinMetadata> {
   constructor(
@@ -16,8 +17,14 @@ export class PMEnvironmentUpdateAction implements IPMACAction<PostmanEnvironment
       throw new Error('pmac workspace id not found')
     }
 
+    const pmacMap = await new PMACEnvironmentGetPMACMapAction(this.pmacWorkspace, this.pmEnvironment).run()
+
+    if (!pmacMap || !pmacMap.pmUID) {
+      throw new Error('pm environment uid not found in pmac workspace')
+    }
+
     const { data: { environment: resetPMEnvironmentRes } } = await this.postmanApi.environments.updateEnvironment(
-      this.pmacWorkspace.pmID,
+      pmacMap.pmUID,
       {
         name: this.pmEnvironment.name,
         values: [],
@@ -25,7 +32,7 @@ export class PMEnvironmentUpdateAction implements IPMACAction<PostmanEnvironment
     )
 
     const { data: { environment: newPMEnvironment } } = await this.postmanApi.environments.updateEnvironment(
-      this.pmacWorkspace.pmID,
+      pmacMap.pmUID,
       {
         name: this.pmEnvironment.name,
         values: this.pmEnvironment.values,
@@ -33,43 +40,5 @@ export class PMEnvironmentUpdateAction implements IPMACAction<PostmanEnvironment
     )
 
     return newPMEnvironment
-
-    //   // Update to reset - because environments update only for "initial value" but not for the current one.
-    //   const { data: { environment } } = await this.postmanApi.environments.updateEnvironment(
-    //     this.pmEnvironmentUid,
-    //     {
-    //       name: this.pmEnvironment.name,
-    //       values: [],
-    //     },
-    //   )
-    // } catch {
-    // When updates fails, try to create the environment
-    // Manage this error somehow
-    // await new PMACEnvironmentCreateAction(
-    //   this.fsWorkspaceResourceManager,
-    //   this.postmanApi,
-    //   this.pmacWorkspace,
-    //   {
-    //     name: this.pmEnvironment.name,
-    //     values: this.pmEnvironment.values,
-    //   }).run()
-    // }
-
-    // If succeed, request the real update
-    // await this.postmanApi.environments.updateEnvironment(
-    //   this.pmEnvironmentUid,
-    //   {
-    //     name: this.pmEnvironment.name,
-    //     values: this.pmEnvironment.values,
-    //   },
-    // )
-
-    // post update
-    // const environment = await new PMEnvironmentPullToPMACAction(
-    //   this.fsWorkspaceResourceManager,
-    //   this.postmanApi,
-    //   this.pmacWorkspace,
-    //   this.pmEnvironmentUid,
-    // ).run()
   }
 }